refactor(backgroundImageList): 파일 위치 이동, 절대 경로 constants 추가
